finally, I fixed it

+++ src/fs/directory.cc
@@ -65,7 +65,7 @@ namespace eclipse {
       FileInfo tmp_file;
       tmp_file.name      = argv[i++];
       tmp_file.hash_key  = atoi(argv[i++]);
-      tmp_file.size      = atoi(argv[i++]);
+      tmp_file.size      = atoll(argv[i++]);
       tmp_file.num_block = atoi(argv[i++]);
       tmp_file.type      = atoi(argv[i++]);
       tmp_file.replica   = atoi(argv[i]);
